Update strategy for to-do list state management
Prior to this change, we use a combination of
`useState` and `useContext` to avoid prop drilling.
While that is valid, it requires invocation of the
context at the top level which in this structure felt
unnecessary and not modular (since the context
is only really required for the to-do list feature).

This commit updates the strategy for the state
management for the to-do list feature to using
`useContext` and `useReducer`.

- All reducer logic can be housed in `taskReducer`
- Initial state and context logic can be housed in
`ToDoContext.js`
- A slight tradeoff with invoking `useReducer` in
`ToDoContainer` instead of a function in
`ToDoContext`. This is to accommodate the
provider value passing into multiple children
within the container.

This should achieve efficient and modular state
management using React hooks without the need
to use Redux (which is too small of a project to
warrant its usage)